Shared toolbar: re-implement PR #16
The change in #16 cannot be merged due to changes in the plugin code.

I've reimplemented the code here and checked the example implementation and it looks good to me. Happy to merge this change or the original if it is re-sync'd with master.

I've made the functionality optional to make it easier for it to be disabled if bugs or issues arise later.

Ping @shihabuddin @hamnis